+++ BE/src/auth/dto/auth-access-token.dto.ts
@@ -1,0 +1,7 @@
+export class AccessTokenDto {
+  accessToken: string;
+
+  constructor(accessToken: string) {
+    this.accessToken = accessToken;
+  }
+}
